+++ application/views/conecte/visualizar_os.php
@@ -75,7 +75,7 @@ $totalProdutos = 0;?>
                                     <tr>
                                         <td>
                                             <strong>Descrição</strong><br> 
-                                            <?php echo $result->descricaoProduto?>
+                                            <?php echo htmlspecialchars_decode($result->descricaoProduto) ?>
                                         </td>
                                     </tr>
                                 
@@ -85,7 +85,7 @@ $totalProdutos = 0;?>
                                     <tr>
                                         <td>
                                             <strong>Defeito</strong><br> 
-                                            <?php echo $result->defeito?>
+                                            <?php echo htmlspecialchars_decode($result->defeito) ?>
                                         </td>
                                     </tr>
                                 <?php }?>
@@ -94,7 +94,7 @@ $totalProdutos = 0;?>
                                         <tr>
                                             <td>
                                                 <strong>Laudo Técnico</strong> <br>
-                                                <?php echo $result->laudoTecnico?>
+                                                <?php echo htmlspecialchars_decode($result->laudoTecnico) ?>
                                             </td>
                                         </tr>
                                 <?php }?>
@@ -103,7 +103,7 @@ $totalProdutos = 0;?>
                                     <tr>
                                         <td>
                                             <strong>Observações</strong> <br>
-                                            <?php echo $result->observacoes?>
+                                            <?php echo htmlspecialchars_decode($result->observacoes) ?>
                                         </td>
                                     </tr>
                                 <?php }?>