+++ application/views/os/adicionarOs.php
@@ -1,6 +1,21 @@
 <link rel="stylesheet" href="<?php echo base_url();?>assets/js/jquery-ui/css/smoothness/jquery-ui-1.9.2.custom.css" />
 <script type="text/javascript" src="<?php echo base_url()?>assets/js/jquery-ui/js/jquery-ui-1.9.2.custom.js"></script>
 <script type="text/javascript" src="<?php echo base_url()?>assets/js/jquery.validate.js"></script>
+
+<link rel="stylesheet" href="<?php echo base_url()?>assets/trumbowyg/ui/trumbowyg.css">
+<script type="text/javascript" src="<?php echo base_url()?>assets/trumbowyg/trumbowyg.js"></script>
+<script type="text/javascript" src="<?php echo base_url()?>assets/trumbowyg/langs/pt_br.js"></script>
+
+<style>
+    .ui-datepicker{
+        z-index: 9999 !important;
+    }
+    .trumbowyg-box {
+        margin-top: 0;
+        margin-bottom: 0;
+    }
+</style>
+
 <div class="row-fluid" style="margin-top:0">
     <div class="span12">
         <div class="widget-box">
@@ -11,9 +26,9 @@
                 <h5>Cadastro de OS</h5>
             </div>
             <div class="widget-content nopadding">
+                <div class="span12" id="divProdutosServicos" style=" margin-left: 0">
                 
 
-                <div class="span12" id="divProdutosServicos" style=" margin-left: 0">
                     <ul class="nav nav-tabs">
                         <li class="active" id="tabDetalhes"><a href="#tab1" data-toggle="tab">Detalhes da OS</a></li>
                     </ul>
@@ -65,27 +80,25 @@
                                     </div>
 
                                     <div class="span12" style="padding: 1%; margin-left: 0">
-
-                                        <div class="span6">
-                                            <label for="descricaoProduto">Descrição Produto/Serviço</label>
-                                            <textarea class="span12" name="descricaoProduto" id="descricaoProduto" cols="30" rows="5"></textarea>
-                                        </div>
-                                        <div class="span6">
-                                            <label for="defeito">Defeito</label>
-                                            <textarea class="span12" name="defeito" id="defeito" cols="30" rows="5"></textarea>
-                                        </div>
-
+                                        <label for="descricaoProduto">Descrição Produto/Serviço</label>
+                                        <textarea class="span12 editor" name="descricaoProduto" id="descricaoProduto" cols="30" rows="5"></textarea>
                                     </div>
+
                                     <div class="span12" style="padding: 1%; margin-left: 0">
-                                        <div class="span6">
-                                            <label for="observacoes">Observações</label>
-                                            <textarea class="span12" name="observacoes" id="observacoes" cols="30" rows="5"></textarea>
-                                        </div>
-                                        <div class="span6">
-                                            <label for="laudoTecnico">Laudo Técnico</label>
-                                            <textarea class="span12" name="laudoTecnico" id="laudoTecnico" cols="30" rows="5"></textarea>
-                                        </div>
+                                        <label for="defeito">Defeito</label>
+                                        <textarea class="span12 editor" name="defeito" id="defeito" cols="30" rows="5"></textarea>
                                     </div>
+
+                                    <div class="span12" style="padding: 1%; margin-left: 0">
+                                        <label for="observacoes">Observações</label>
+                                        <textarea class="span12 editor" name="observacoes" id="observacoes" cols="30" rows="5"></textarea>
+                                    </div>
+
+                                    <div class="span12" style="padding: 1%; margin-left: 0">
+                                        <label for="laudoTecnico">Laudo Técnico</label>
+                                        <textarea class="span12 editor" name="laudoTecnico" id="laudoTecnico" cols="30" rows="5"></textarea>
+                                    </div>
+
                                     <div class="span12" style="padding: 1%; margin-left: 0">
                                         <div class="span6 offset3" style="text-align: center">
                                             <button class="btn btn-success" id="btnContinuar"><i class="icon-share-alt icon-white"></i> Continuar</button>
@@ -164,7 +177,10 @@ $(document).ready(function(){
        });
 
     $(".datepicker" ).datepicker({ dateFormat: 'dd/mm/yy' });
-   
+
+    $('.editor').trumbowyg({
+        lang: 'pt_br'
+    });
 });
 
 </script>