+++ application/views/os/imprimirOs.php
@@ -107,7 +107,7 @@ $totalProdutos = 0;?>
                             <tr>
                                 <td colspan="3">
                                 <b>DESCRIÇÃO: </b>
-                                <?php echo $result->descricaoProduto ?>
+                                <?php echo htmlspecialchars_decode($result->descricaoProduto) ?>
                                 </td>
                             </tr>
                             <?php }?>
@@ -116,7 +116,7 @@ $totalProdutos = 0;?>
                             <tr>
                                 <td colspan="3">
                                 <b>DEFEITO APRESENTADO: </b>
-                                <?php echo $result->defeito?>
+                                <?php echo htmlspecialchars_decode($result->defeito) ?>
                                 </td>
                             </tr>
                             <?php }?>
@@ -125,7 +125,7 @@ $totalProdutos = 0;?>
                             <tr>
                                 <td colspan="3">
                                 <b>OBSERVAÇÕES: </b>
-                                <?php echo $result->observacoes?>
+                                <?php echo htmlspecialchars_decode($result->observacoes) ?>
                                 </td>
                             </tr>
                             <?php }?>
@@ -134,7 +134,7 @@ $totalProdutos = 0;?>
                             <tr>
                                 <td colspan="3">
                                 <b>LAUDO TÉCNICO: </b>
-                                <?php echo $result->laudoTecnico?>
+                                <?php echo htmlspecialchars_decode($result->laudoTecnico) ?>
                                 </td>
                             </tr>
                             <?php }?>