+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Todas as alterações serão documentadas neste arquivo
 Formato baseado em [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 e [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2] - 2019-02-18
+### Added
+- Implementado editor WYSIWYG [Trumbowyg](https://github.com/Alex-D/Trumbowyg) nos campos (Ordem de Serviço)
+    - Descrição Produto/Serviço
+    - Defeito
+    - Observações
+    - Laudo Técnico
+
 ## [3.1.16] - 2019-01-14
 ### Added
 - Adicionando e-mail do cliente na impressão da OS. [@RamonSilva20](https://github.com/RamonSilva20).